+++ angular-workspace/projects/ni/nimble-angular/table-column/icon/nimble-table-column-icon.directive.ts
@@ -18,8 +18,6 @@ export class NimbleTableColumnIconDirective extends NimbleTableColumnBaseDirecti
         return this.elementRef.nativeElement.fieldName;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('field-name') public set fieldName(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'fieldName', value);
     }
@@ -28,8 +26,6 @@ export class NimbleTableColumnIconDirective extends NimbleTableColumnBaseDirecti
         return this.elementRef.nativeElement.keyType;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('key-type') public set keyType(value: MappingKeyType) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'keyType', value);
     }
@@ -38,8 +34,6 @@ export class NimbleTableColumnIconDirective extends NimbleTableColumnBaseDirecti
         return this.elementRef.nativeElement.fractionalWidth;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('fractional-width') public set fractionalWidth(value: NumberValueOrAttribute | null | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'fractionalWidth', toNullableNumberProperty(value));
     }
@@ -48,8 +42,6 @@ export class NimbleTableColumnIconDirective extends NimbleTableColumnBaseDirecti
         return this.elementRef.nativeElement.minPixelWidth;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('min-pixel-width') public set minPixelWidth(value: NumberValueOrAttribute | null | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'minPixelWidth', toNullableNumberProperty(value));
     }
@@ -58,8 +50,6 @@ export class NimbleTableColumnIconDirective extends NimbleTableColumnBaseDirecti
         return this.elementRef.nativeElement.groupIndex;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('group-index') public set groupIndex(value: NumberValueOrAttribute | null | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupIndex', toNullableNumberProperty(value));
     }
@@ -68,8 +58,6 @@ export class NimbleTableColumnIconDirective extends NimbleTableColumnBaseDirecti
         return this.elementRef.nativeElement.groupingDisabled;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('grouping-disabled') public set groupingDisabled(value: BooleanValueOrAttribute) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'groupingDisabled', toBooleanProperty(value));
     }