+++ packages/nimble-components/src/utilities/unit-format/manually-translated-unit-scale/base/unit-translation.ts
@@ -1,0 +1,10 @@
+/**
+ * Representations of a base unit in a particular language
+ */
+export class UnitTranslation {
+    public constructor(
+        public readonly singular: string,
+        public readonly plural: string,
+        public readonly symbol: string
+    ) {}
+}